refac: move include/{db,nc,pl} into src/overlays, rename include/an to include/core1, move header files into 'correct' directory where applicable
