+++ include/bs/kazow.h
@@ -3,7 +3,7 @@
 
 #include <ultra64.h>
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/bounce.h"
 #include "ba/physics.h"
