+++ include/bs/kazpaddle.h
@@ -6,7 +6,7 @@
 #include "bs/drone.h"
 #include "ba/playerstate.h"
 #include "bs/kaz.h"
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/input.h"
 #include "ba/key.h"
