+++ include/bs/kazshock.h
@@ -3,7 +3,7 @@
 
 #include <ultra64.h>
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/input.h"
 #include "ba/key.h"
