+++ include/bs/kazstilt.h
@@ -3,7 +3,7 @@
 
 #include <ultra64.h>
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/playerstate.h"
 #include "ba/anim.h"
 #include "ba/yaw.h"
