+++ include/bs/mumattack.h
@@ -1,7 +1,7 @@
 #ifndef __BS_MUMATTACK_H__
 #define __BS_MUMATTACK_H__
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/flag.h"
 #include "ba/key.h"
