+++ include/bs/mummove.h
@@ -1,7 +1,7 @@
 #ifndef __BS_MUMMOVE_H__
 #define __BS_MUMMOVE_H__
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/drone.h"
 #include "ba/dust.h"
