+++ src/core1/mlmtx.c
@@ -1,4 +1,4 @@
-#include "ml/mtx.h"
+#include "core1/mlmtx.h"
 
 MtxF D_8007B4F0[0x1F]; // mtx_buffer
 MtxF* s_mtx_stack;
