+++ src/core2/1ECBCC0.c
@@ -1,5 +1,5 @@
 #include <ultra64.h>
-#include "ml/mtx.h"
+#include "core1/mlmtx.h"
 
 #pragma GLOBAL_ASM("asm/nonmatchings/core2/1ECBCC0/func_800F23D0.s")
