+++ src/core2/baanim.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/physics.h"
