+++ src/overlays/bsbbuster/bsbbuster.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/assets.h"
 #include "ba/dust.h"
