+++ src/overlays/bsbflip/bsbflip.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/anim.h"
 #include "ba/playerstate.h"
 #include "ba/physics.h"
