+++ src/overlays/bstwirl/bstwirl.c
@@ -1,7 +1,7 @@
 #include "common.h"
 
 #include "ba/anim.h"
-#include "an/anctrl.h"
+#include "core2/anctrl.h"
 #include "ba/assets.h"
 #include "ba/physics.h"
 #include "bs.h"
