+++ src/overlays/nc/badie.h
